Add files via upload
This is a music store manager program in C++ with mysql database that can create order, calculate bill, add new music in database, edit the music description, and delete any item also it shows the total music in stock, and it can find the specific music. it stores all data in MySQL database.